command tutorial stored in struct

+++ cmd/tutorials/main.go
@@ -2,24 +2,30 @@ package main
 
 import (
 	"fmt"
-	"image"
 	"kaellybot/image-renderer/pkg/automations"
+	"kaellybot/image-renderer/pkg/constants"
 	"kaellybot/image-renderer/pkg/images"
 	"log"
 	"os"
 	"path/filepath"
 	"sync"
-	"time"
 
 	amqp "github.com/kaellybot/kaelly-amqp"
 )
 
 func main() {
-	bounds := image.Rect(350, 200, 1000, 1010)
-	duration := 10 * time.Second
+	commandName := "about"
+	locale := amqp.Language_FR
 	fps := 15
+
+	command, found := constants.GetCommands()[commandName]
+	if !found {
+		log.Printf("command %v not found\n", commandName)
+		return
+	}
+
 	outputDir := filepath.Join(os.TempDir(), "kaelly_frames")
-	outputWEBP := "out.webp"
+	outputWEBP := fmt.Sprintf("%v_%v.webp", command.Name, locale)
 
 	os.Mkdir(outputDir, 0755)
 	defer os.RemoveAll(outputDir)
@@ -33,7 +39,7 @@ func main() {
 	wg.Add(2)
 	go func() {
 		log.Println("Recording...")
-		if err := images.RecordScreen(outputDir, duration, fps, bounds); err != nil {
+		if err := images.RecordScreen(outputDir, command.Duration, fps, command.Bounds); err != nil {
 			log.Printf("ffmpeg recording failed: %v\n", err)
 		}
 		wg.Done()
@@ -41,7 +47,7 @@ func main() {
 
 	go func() {
 		log.Println("Automating...")
-		if err := automations.RunCommandTutorial("about", amqp.Language_FR); err != nil {
+		if err := automations.RunCommandTutorial(command, amqp.Language_FR); err != nil {
 			log.Printf("ffmpeg recording failed: %v\n", err)
 		}
 		wg.Done()
